+++ Content.Server/_CE/ZLevels/Mapping/Commands/CEInitializeZNetworkCommand.cs
@@ -23,7 +23,7 @@ public sealed partial class CEInitializeZNetworkCommand : LocalizedEntityCommand
     public override CompletionResult GetCompletion(IConsoleShell shell, string[] args)
     {
         var options = new List<CompletionOption>();
-        var query = _entities.EntityQueryEnumerator<CEZLevelsNetworkComponent, MetaDataComponent>();
+        var query = _entities.EntityQueryEnumerator<CEZMapNetworkComponent, MetaDataComponent>();
         while (query.MoveNext(out var uid, out var zLevelComp, out var meta))
         {
             options.Add(new CompletionOption(_entities.GetNetEntity(uid).ToString(), meta.EntityName));
@@ -50,7 +50,7 @@ public sealed partial class CEInitializeZNetworkCommand : LocalizedEntityCommand
             return;
         }
 
-        if (!_entities.TryGetComponent<CEZLevelsNetworkComponent>(target, out var levelComp))
+        if (!_entities.TryGetComponent<CEZMapNetworkComponent>(target, out var levelComp))
         {
             shell.WriteError($"Target entity doesnt have CEZLevelsNetworkComponent {args[1]}");
             return;
