+++ Content.Shared/_CE/ZLevels/Core/Components/CEZMapComponent.cs
@@ -11,7 +11,7 @@ namespace Content.Shared._CE.ZLevels.Core.Components;
 /// Automatically added to the map when it appears in zLevelNetwork.
 /// </summary>
 [RegisterComponent, NetworkedComponent, AutoGenerateComponentState, UnsavedComponent]
-public sealed partial class CEZLevelMapComponent : Component
+public sealed partial class CEZMapComponent : Component
 {
     [ViewVariables, AutoNetworkedField]
     public EntityUid NetworkUid;
