--- Content.Shared/_CE/ZLevels/Core/EntitySystems/CESharedZLevelsSystem.Activation.API.cs
@@ -1,66 +1,0 @@
-/*
- * This file is sublicensed under MIT License
- * https://github.com/space-wizards/space-station-14/blob/master/LICENSE.TXT
- */
-
-using Content.Shared._CE.ZLevels.Core.Components;
-using JetBrains.Annotations;
-using Robust.Shared.Physics;
-
-namespace Content.Shared._CE.ZLevels.Core.EntitySystems;
-
-public abstract partial class CESharedZLevelsSystem
-{
-    [PublicAPI]
-    public void WakeBody(Entity<CEZPhysicsComponent?> entity)
-    {
-        if (!Resolve(entity, ref entity.Comp, false))
-            return;
-
-        if (_activeBodies.Contains(entity))
-            return;
-
-        entity.Comp.Sleeping = false;
-        entity.Comp.SleepTimer = 0f;
-
-        _activeBodies.Add(entity);
-    }
-
-    [PublicAPI]
-    public void SleepBody(Entity<CEZPhysicsComponent?> entity)
-    {
-        if (!Resolve(entity, ref entity.Comp, false))
-            return;
-
-        entity.Comp.Sleeping = true;
-        entity.Comp.SleepTimer = 0f;
-
-        _activeBodies.Remove(entity);
-    }
-
-    [PublicAPI]
-    public void RefreshBody(Entity<CEZPhysicsComponent> entity)
-    {
-        if (TerminatingOrDeleted(entity))
-        {
-            SleepBody((entity, entity));
-            return;
-        }
-
-        var transform = Transform(entity);
-        var parent = transform.ParentUid;
-
-        var onMap = parent == transform.GridUid || parent == transform.MapUid;
-
-        if (!onMap
-            || transform.Anchored
-            || _physicsQuery.TryComp(entity, out var physics)
-            && physics.BodyType == BodyType.Static)
-        {
-            SleepBody((entity, entity));
-            return;
-        }
-
-        WakeBody((entity, entity));
-    }
-}
