+++ Content.Shared/_CE/ZLevels/Core/EntitySystems/CESharedZLevelsSystem.Activation.cs
@@ -4,6 +4,8 @@
  */
 
 using Content.Shared._CE.ZLevels.Core.Components;
+using JetBrains.Annotations;
+using Robust.Shared.Physics;
 using Robust.Shared.Physics.Events;
 
 namespace Content.Shared._CE.ZLevels.Core.EntitySystems;
@@ -70,5 +72,58 @@ public abstract partial class CESharedZLevelsSystem
         {
             SetZPosition((entity, entity), 0);
         }
+    }
+
+    [PublicAPI]
+    public void WakeBody(Entity<CEZPhysicsComponent?> entity)
+    {
+        if (!Resolve(entity, ref entity.Comp, false))
+            return;
+
+        if (_activeBodies.Contains(entity))
+            return;
+
+        entity.Comp.Sleeping = false;
+        entity.Comp.SleepTimer = 0f;
+
+        _activeBodies.Add(entity);
+    }
+
+    [PublicAPI]
+    public void SleepBody(Entity<CEZPhysicsComponent?> entity)
+    {
+        if (!Resolve(entity, ref entity.Comp, false))
+            return;
+
+        entity.Comp.Sleeping = true;
+        entity.Comp.SleepTimer = 0f;
+
+        _activeBodies.Remove(entity);
+    }
+
+    [PublicAPI]
+    public void RefreshBody(Entity<CEZPhysicsComponent> entity)
+    {
+        if (TerminatingOrDeleted(entity))
+        {
+            SleepBody((entity, entity));
+            return;
+        }
+
+        var transform = Transform(entity);
+        var parent = transform.ParentUid;
+
+        var onMap = parent == transform.GridUid || parent == transform.MapUid;
+
+        if (!onMap
+            || transform.Anchored
+            || _physicsQuery.TryComp(entity, out var physics)
+            && physics.BodyType == BodyType.Static)
+        {
+            SleepBody((entity, entity));
+            return;
+        }
+
+        WakeBody((entity, entity));
     }
 }
