+++ Content.Shared/_CE/ZLevels/Core/EntitySystems/CESharedZLevelsSystem.Grids.cs
@@ -11,8 +11,8 @@ namespace Content.Shared._CE.ZLevels.Core.EntitySystems;
 
 public abstract partial class CESharedZLevelsSystem
 {
-    [Dependency] protected EntityQuery<CEZGridComponent> _zgridQuery = default!;
-    [Dependency] protected EntityQuery<CEZGridNetworkComponent> _zgridNetworkQuery = default!;
+    [Dependency] private EntityQuery<CEZGridComponent> _zGridQuery = default!;
+    [Dependency] private EntityQuery<CEZGridNetworkComponent> _zGridNetworkQuery = default!;
 
     /// <summary>
     /// Returns the z-level depth of the map containing the given grid,
@@ -45,11 +45,11 @@ public abstract partial class CESharedZLevelsSystem
     {
         network = default;
 
-        if (!_zgridQuery.TryComp(grid, out var zGridComp) || zGridComp.NetworkId == string.Empty)
+        if (!_zGridQuery.TryComp(grid, out var zGridComp) || zGridComp.NetworkId == string.Empty)
             return false;
 
         // Fast path
-        if (zGridComp.Network.IsValid() && _zgridNetworkQuery.TryComp(zGridComp.Network, out var cached))
+        if (zGridComp.Network.IsValid() && _zGridNetworkQuery.TryComp(zGridComp.Network, out var cached))
         {
             network = (zGridComp.Network, cached);
             return true;
@@ -71,10 +71,15 @@ public abstract partial class CESharedZLevelsSystem
 }
 
 /// <summary>
+/// Called on ZLevel grid Network Entity, when grid added or removed from network
+/// </summary>
+public sealed class CEZLevelGridNetworkUpdatedEvent : EntityEventArgs;
+
+/// <summary>
 /// Directed at a grid entity when it is added to a z-grid network by <see cref="CEZGridConnectorSystem"/>.
 /// </summary>
 [ByRefEvent]
-public readonly struct CEGridLinkedEvent(EntityUid network)
+public readonly struct CEGridAddedIntoZNetworkEvent(EntityUid network)
 {
     public readonly EntityUid Network = network;
 }
@@ -84,7 +89,7 @@ public readonly struct CEGridLinkedEvent(EntityUid network)
 /// either by the recalculator or by external network deletion.
 /// </summary>
 [ByRefEvent]
-public readonly struct CEGridUnlinkedEvent(EntityUid network)
+public readonly struct CEGridRemovedFromZNetworkEvent(EntityUid network)
 {
     public readonly EntityUid Network = network;
 }
