+++ Content.Shared/_CE/ZLevels/Core/EntitySystems/CESharedZLevelsSystem.cs
@@ -35,8 +35,6 @@ public abstract partial class CESharedZLevelsSystem : EntitySystem
     [Dependency] private SharedMapSystem _map = null!;
     [Dependency] private SharedPopupSystem _popup = null!;
 
-    #region Queries
-
     [Dependency] protected EntityQuery<CEZPhysicsComponent> ZPhysicsQuery = default!;
 
     [Dependency] private EntityQuery<MapComponent> _mapQuery = default!;
@@ -44,11 +42,9 @@ public abstract partial class CESharedZLevelsSystem : EntitySystem
     [Dependency] private EntityQuery<PhysicsComponent> _physicsQuery = default!;
     [Dependency] private EntityQuery<TransformComponent> _transformQuery = default!;
 
-    [Dependency] private EntityQuery<CEZLevelMapComponent> _zMapQuery = default!;
-    [Dependency] private EntityQuery<CEZLevelsNetworkComponent> _zNetworkQuery = default!;
+    [Dependency] private EntityQuery<CEZMapComponent> _zMapQuery = default!;
+    [Dependency] private EntityQuery<CEZMapNetworkComponent> _zNetworkQuery = default!;
     [Dependency] private EntityQuery<CEZLevelHighGroundComponent> _zHighGroundQuery = default!;
-
-    #endregion
 
     private bool _clientSimulation;
     private TimeSpan _fixedTimestep;
@@ -64,180 +60,5 @@ public abstract partial class CESharedZLevelsSystem : EntitySystem
         InitializeCacheHooks();
         InitializeMovement();
         InitializeView();
-    }
-
-    /// <summary>
-    /// Checks whether the map is in the zLevels network. If so, returns true and the current depth + Entity of the current zLevels network.
-    /// </summary>
-    [PublicAPI]
-    public bool TryGetZNetwork(EntityUid mapUid, out Entity<CEZLevelsNetworkComponent> zLevel)
-    {
-        zLevel = default;
-        if (!_zMapQuery.TryComp(mapUid, out var zMap))
-            return false;
-
-        var networkUid = zMap.NetworkUid;
-        if (TerminatingOrDeleted(networkUid))
-        {
-            Log.Warning($"Trying access to terminated z-network, map: {mapUid}, outdated network uid: {networkUid}");
-            return false;
-        }
-
-        if (!_zNetworkQuery.TryComp(networkUid, out var zNetworkComponent))
-        {
-            Log.Warning($"Trying access to z-network without component??? WHY?! map: {mapUid}, network uid: {networkUid}");
-            return false;
-        }
-
-        zLevel = (networkUid, zNetworkComponent);
-        return true;
-    }
-
-    /// <summary>
-    /// Returns the inclusive depth range of the z-network the given map belongs to
-    /// (<c>minDepth</c>..<c>maxDepth</c>). False if the map is not in a z-network.
-    /// </summary>
-    [PublicAPI]
-    public bool TryGetDepthBounds(EntityUid mapUid, out int minDepth, out int maxDepth)
-    {
-        minDepth = 0;
-        maxDepth = 0;
-        if (!TryGetZNetwork(mapUid, out var network))
-            return false;
-
-        minDepth = network.Comp.SortedMin;
-        maxDepth = network.Comp.SortedMax;
-        return true;
-    }
-
-    [PublicAPI]
-    public bool TryMapOffset(Entity<CEZLevelMapComponent?> entity, int offset, out Entity<CEZLevelMapComponent> output)
-    {
-        return TryMapOffset(entity, offset, out output, out _);
-    }
-
-    [PublicAPI]
-    public bool TryMapOffset(Entity<CEZLevelMapComponent?> entity, int offset, out Entity<CEZLevelMapComponent> output, [NotNullWhen(true)] out MapComponent? mapComponent)
-    {
-        output = default;
-        mapComponent = null;
-
-        if (MapOffset(entity, offset) is not { } result)
-            return false;
-
-        if (!TryComp(result, out mapComponent))
-            return false;
-
-        output = result;
-        return true;
-    }
-
-    [PublicAPI]
-    public Entity<CEZLevelMapComponent>? MapOffset(Entity<CEZLevelMapComponent?> entity, int offset)
-    {
-        if (!Resolve(entity, ref entity.Comp, false))
-            return null;
-
-        var target = offset switch
-        {
-             1 => entity.Comp.MapAbove,
-            -1 => entity.Comp.MapBelow,
-             _ => null,
-        };
-
-        if (target is not null && _zMapQuery.TryComp(target.Value, out var component))
-            return (target.Value, component);
-
-        if (!_zNetworkQuery.TryComp(entity.Comp.NetworkUid, out var network))
-            return null;
-
-        if (!network.ZLevels.TryGetValue(entity.Comp.Depth + offset, out var targetId))
-            return null;
-
-        return _zMapQuery.TryComp(targetId, out var comp)
-            ? (targetId.Value, comp)
-            : null;
-    }
-
-    [PublicAPI]
-    public bool TryMapUp(Entity<CEZLevelMapComponent?> inputMapUid, out Entity<CEZLevelMapComponent> aboveMapUid) =>
-        TryMapOffset(inputMapUid, 1, out aboveMapUid);
-
-    [PublicAPI]
-    public bool TryMapDown(Entity<CEZLevelMapComponent?> inputMapUid, out Entity<CEZLevelMapComponent> belowMapUid) =>
-        TryMapOffset(inputMapUid, -1, out belowMapUid);
-
-    /// <summary>
-    /// Returns a list of all maps above the specified map. The closest map at the top is returned first.
-    /// </summary>
-    [PublicAPI]
-    public List<EntityUid> GetAllMapsAbove(Entity<CEZLevelMapComponent> entity)
-    {
-        if (!_zNetworkQuery.TryComp(entity.Comp.NetworkUid, out var network) || entity.Comp.Depth >= network.SortedMax)
-            return new List<EntityUid>();
-
-        var startIndex = entity.Comp.Depth < network.SortedMin ? 0 : entity.Comp.Depth - network.SortedMin + 1;
-        var estimatedCount = network.SortedZLevels.Count - startIndex;
-        var result = new List<EntityUid>(estimatedCount);
-        var zLevels = network.SortedZLevels;
-
-        for (var i = startIndex; i < zLevels.Count; i++)
-        {
-            var uid = zLevels[i];
-            if (uid == EntityUid.Invalid)
-                continue;
-
-            if (_zMapQuery.HasComp(uid))
-                result.Add(uid);
-        }
-
-        return result;
-    }
-
-    [PublicAPI]
-    public void GetAllMapsAbove(Entity<CEZLevelMapComponent> entity, List<EntityUid> result)
-    {
-        result.Clear();
-
-        if (!_zNetworkQuery.TryComp(entity.Comp.NetworkUid, out var network) || entity.Comp.Depth >= network.SortedMax)
-            return;
-
-        var startIndex = entity.Comp.Depth < network.SortedMin ? 0 : entity.Comp.Depth - network.SortedMin + 1;
-        var zLevels = network.SortedZLevels;
-
-        for (var i = startIndex; i < zLevels.Count; i++)
-        {
-            var uid = zLevels[i];
-
-            if (uid.IsValid() && _zMapQuery.HasComp(uid))
-                result.Add(uid);
-        }
-    }
-
-    /// <summary>
-    /// Returns a list of all maps below the specified map. The closest map at the bottom is returned first.
-    /// </summary>
-    [PublicAPI]
-    public List<EntityUid> GetAllMapsBelow(Entity<CEZLevelMapComponent> entity)
-    {
-        if (!_zNetworkQuery.TryComp(entity.Comp.NetworkUid, out var network) || entity.Comp.Depth <= network.SortedMin)
-            return new List<EntityUid>();
-
-        var endIndex = entity.Comp.Depth - network.SortedMin;
-        var result = new List<EntityUid>(endIndex);
-        var zLevels = network.SortedZLevels;
-
-        // Iterate backwards for nearest-first ordering.
-        for (var i = endIndex - 1; i >= 0; i--)
-        {
-            var uid = zLevels[i];
-            if (uid == EntityUid.Invalid)
-                continue;
-
-            if (_zMapQuery.HasComp(uid))
-                result.Add(uid);
-        }
-
-        return result;
     }
 }
